Refactored Uncouple Recouple so we only have 1 version
