+++ src/pages/test-report/cat-be/test-report.cat-be.page.html
@@ -103,7 +103,7 @@
         </ion-row>
         <ion-row>
           <ion-col>
-           <uncouple-recouple></uncouple-recouple>
+           <uncouple-recouple [category]="'B+E'"></uncouple-recouple>
           </ion-col>
         </ion-row>
         <ion-row>
