--- src/pages/test-report/cat-c/components/uncouple-recouple/uncouple-recouple.scss
@@ -1,48 +1,0 @@
-uncouple-recouple {
-  competency-button {
-    position: relative;
-  }
-
-  competency-button.uncouple-recouple-tick {
-
-    .competency-button {
-      padding: 1px 7px 2px;
-
-      .tickwrapper {
-        margin: auto;
-        display: flex;
-        justify-content: center;
-      }
-    }
-
-    &.checked .competency-button {
-      border: 2px solid map-get($colors-gds, "gds-dark-green");
-    }
-  }
-
-  competency-button.uncouple-recouple-competency {
-    display: flex;       // this styling is neccessary to ensure z-index
-    flex-flow: column;   // works on competency buttons
-    .competency-button {
-      margin-left: 8px;
-
-      padding-left: 5px;
-      padding-right: 5px;
-
-      driving-faults-badge {
-        width: 30px;
-      }
-
-      .fault-wrapper {
-        display: flex;
-        width: 30px;
-        justify-content: flex-end;
-        align-items: center;
-
-        dangerous-fault-badge {
-          margin-left: -22px;
-        }
-      }
-    }
-  }
-}
