+++ src/pages/test-report/cat-c/test-report.cat-c.page.html
@@ -108,7 +108,7 @@
         </ion-row>
         <ion-row [ngStyle]="{'visibility': showUncoupleRecouple() ? 'visible' : 'hidden' }">
           <ion-col>
-           <uncouple-recouple></uncouple-recouple>
+           <uncouple-recouple [category]="pageState.testCategory$ | async"></uncouple-recouple>
           </ion-col>
         </ion-row>
         <ion-row *ngIf="showUncoupleRecouple()">
