+++ src/pages/test-report/cat-d/test-report.cat-d.page.html
@@ -107,7 +107,7 @@
         </ion-row>
         <ion-row [ngStyle]="{'visibility': showUncoupleRecouple() ? 'visible' : 'hidden' }">
           <ion-col>
-            <uncouple-recouple-cat-d></uncouple-recouple-cat-d>
+            <uncouple-recouple [category]="pageState.testCategory$ | async"></uncouple-recouple>
           </ion-col>
         </ion-row>
         <ion-row *ngIf="showUncoupleRecouple()">
@@ -118,7 +118,7 @@
         <ion-row>
           <ion-col>
             <!-- PLACEHOLDER FOR DOOR EXERCISE-->
-            <uncouple-recouple-cat-d></uncouple-recouple-cat-d>
+            <uncouple-recouple [category]="pageState.testCategory$ | async"></uncouple-recouple>
           </ion-col>
         </ion-row>
         <ion-row >
